ISSUE #252
* Extracted two new test cases to cover testing with environment variables and parameterized input.
* Refactored MinosConfigCreation into setUp.
