+++ tests/test_common/test_configuration/test_config.py
@@ -5,11 +5,7 @@ This file is part of minos framework.
 
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
-import os
 import unittest
-from unittest import (
-    mock,
-)
 
 from minos.common import (
     MinosConfig,
@@ -25,23 +21,22 @@ from tests.utils import (
 class TestMinosConfig(unittest.TestCase):
     def setUp(self) -> None:
         self.config_file_path = BASE_PATH / "test_config.yml"
+        self.config = MinosConfig(path=self.config_file_path)
 
     def test_config_ini_fail(self):
         with self.assertRaises(MinosConfigException):
             MinosConfig(path=BASE_PATH / "test_fail_config.yaml")
 
     def test_cast_path(self):
-        config = MinosConfig(path=str(self.config_file_path))
-        self.assertEqual(self.config_file_path, config._path)
+        config_path = self.config._path
+        self.assertEqual(self.config_file_path, config_path)
 
     def test_config_service(self):
-        config = MinosConfig(path=self.config_file_path)
-        service = config.service
+        service = self.config.service
         self.assertEqual("Order", service.name)
 
     def test_config_rest(self):
-        config = MinosConfig(path=self.config_file_path)
-        rest = config.rest
+        rest = self.config.rest
 
         broker = rest.broker
         self.assertEqual("localhost", broker.host)
@@ -51,8 +46,7 @@ class TestMinosConfig(unittest.TestCase):
         self.assertEqual("AddOrder", endpoints[0].name)
 
     def test_config_events(self):
-        config = MinosConfig(path=self.config_file_path)
-        events = config.events
+        events = self.config.events
         broker = events.broker
         self.assertEqual("localhost", broker.host)
         self.assertEqual(9092, broker.port)
@@ -117,29 +111,6 @@ class TestMinosConfig(unittest.TestCase):
         self.assertEqual("localhost", snapshot.host)
         self.assertEqual(5432, snapshot.port)
 
-    @mock.patch.dict(os.environ, {"MINOS_REPOSITORY_DATABASE": "foo"})
-    def test_overwrite_with_environment(self):
-        config = MinosConfig(path=self.config_file_path)
-        repository = config.repository
-        self.assertEqual("foo", repository.database)
-
-    @mock.patch.dict(os.environ, {"MINOS_REPOSITORY_DATABASE": "foo"})
-    def test_overwrite_with_environment_false(self):
-        config = MinosConfig(path=self.config_file_path, with_environment=False)
-        repository = config.repository
-        self.assertEqual("order_db", repository.database)
-
-    def test_overwrite_with_parameter(self):
-        config = MinosConfig(path=self.config_file_path, repository_database="foo")
-        repository = config.repository
-        self.assertEqual("foo", repository.database)
-
-    @mock.patch.dict(os.environ, {"MINOS_REPOSITORY_DATABASE": "foo"})
-    def test_overwrite_with_parameter_priority(self):
-        config = MinosConfig(path=self.config_file_path, repository_database="bar")
-        repository = config.repository
-        self.assertEqual("bar", repository.database)
-
     def test_get_default_default(self):
         with MinosConfig(path=self.config_file_path) as config:
             self.assertEqual(config, MinosConfigAbstract.get_default())
@@ -149,7 +120,3 @@ class TestMinosConfig(unittest.TestCase):
             with MinosConfig(path=self.config_file_path):
                 with MinosConfig(path=self.config_file_path):
                     pass
-
-
-if __name__ == "__main__":
-    unittest.main()
